scripts/cf2png-gpu-requirements.txt
pypng>=0.0.18
pycuda>=2016.1.2
